requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
click==8.1.3
    # via flask
flask==2.2.3
    # via -r requirements.in
importlib-metadata==6.3.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
pyaml==21.10.1
    # via pymeterbus
pycryptodome==3.17
    # via pymeterbus
pymeterbus==0.8.3
    # via -r requirements.in
pyserial==3.5
    # via pymeterbus
pyyaml==6.0
    # via pyaml
simplejson==3.19.1
    # via pymeterbus
werkzeug==2.2.3
    # via flask
zipp==3.15.0
    # via importlib-metadata
